refactor(routes/models): extract list_models aggregation to models_service.list_all
P3-routers §J step 6. Move the three-source /api/models catalog aggregation
(registry rows + update_available recompute, genuine-remote upstream
advertisements with FLM-skip + operator filters, host-FLM installed-tag
injection) out of the route handler into hal0.services.models_service.list_all.
The route becomes a request→service→envelope shell wiring app.state.
apply_filters moved into the service; route import dropped.

models.py: 1163 -> 987.
